+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Nodemailer `smtpTransport.SmtpOptions` config setting.
+
+### Deleted
+- `smtpPort` config setting.
+- `smtpHost` config setting.
+
 ## [0.4.0] - 2017-10-6
 ### Added
 - AJAX data request support